+++ AndroidBLE/app/src/main/java/com/pascaldornfeld/gsdble/overview/components/CharaDataStatsBySecondValue.kt
@@ -6,27 +6,27 @@ import kotlin.math.sqrt
 
 class CharaDataStatsBySecondValue(
     timestamp: Long,
-    averageGraph: GraphFragment<Float>?,
-    deviationGraph: GraphFragment<Float>?
+    averageGraph: GraphFragment<Double>?,
+    deviationGraph: GraphFragment<Double>?
 ) : CharaDataStatsCalculatorAsyncTask(timestamp, averageGraph, deviationGraph) {
 
-    override fun doInBackground(vararg params: Array<Pair<Long, Long>>?): Float? {
+    override fun doInBackground(vararg params: Array<Pair<Long, Long>>?): Double? {
         assert(params.isNotEmpty())
         val inputArray = params[0]
 
         if (inputArray != null && inputArray.size >= 2) {
             // calculate average
             val avgTime = inputArray.map { it.second }.average()
-            publishProgress(avgTime.toFloat())
+            publishProgress(avgTime)
             if (isCancelled) return null
 
             // calculate variance
-            var sum = 0.0f
-            inputArray.forEach { sum += (it.second - avgTime).pow(2).toFloat() }
+            var sum = 0.0
+            inputArray.forEach { sum += (it.second.toDouble() - avgTime).pow(2) }
 
             // return deviation
             // sum is based on inputArray.size elements. The formula takes this amount -1. so we got inputArray.size -1
-            return sqrt((sum / (inputArray.size - 1).toFloat()))
+            return sqrt((sum / (inputArray.size - 1)))
         } else {
             publishProgress(null)
             return null
